+++ bindings/virtio/tscclock.c
@@ -210,7 +210,7 @@ int tscclock_init(void) {
     i8254_delay(100000);
     tsc_freq = (cpu_rdtsc() - tsc_base) * 10;
     log(INFO, "Solo5: Clock source: TSC, frequency estimate is %lu Hz\n",
-        (unsigned long long)tsc_freq);
+        tsc_freq);
 
     /*
      * Calculate TSC scaling multiplier.
